Add support for event driven traversal and display
The map of slides comes from a string (as described in README).
That's parsed to create a 2d array. A SlideMap objects encapsulates
that array with next, nextX, nextY, prevX, prevY, reset and find
methods that allow traversing the map.

When the current index into the map (the model) is changed an event
is trigged. A view listens for that event and does the necessary.

There's _some_ Backbone being used here, but not much yet as I wanted
to be more direct (for now).

At this point the necessary is some console logging, some #hash
updating, and very limited dom updating.

The next iteration the map will come from a tiddler and will
indicate tiddlers to load.

Might be nice to have a 'g' event which pops a modal used to
'go' to a particular slide (by index or title).

As noted in the comments some lines of code could be saved in
next* and prev* methods by doing some COMPUTER SCIENCE, but I
wasn't in that kind of mood.

Some bind() going on in this code, thanks to @bengillies and @FND.