Add support for dynamic modification of CSS
The reason for doing this is to facillitate some reasonable defaults
on font sizes and list position.

cssCleanup() creates (or modifies) a stylesheet which sets the font-size
for h1-3 and li, and sets the position of a ul. If an img is a general
sibling of the ul, then the ul is positioned further left, to make
room for the image (which always floats right).

At the moment the CSS is recalculated at:

* startup
* slide change
* window resize
* keypress 'c'

This is probably overkill, but we'll see. 'slide change' is only
there because it makes the lists center better, but that's being
a bit oversensitive?

The changes assume that all you'll ever do is a <ul> and an <img>
which is not likely to be true, so will need to adjust accordingly
as more experience with content is gained.

Also:

* basic CSS (colors, gradient)
* a slide counter

BTW: The CSS is done as a stylesheet rather than just doing $(selector).css..
because the stuff being styled doesn't exist yet.

Next: reading in some tiddlers.